# [ENH]improve conditional join performance