@ericmjl 2024-11-22
@samukweku, I am not 100% sure of the changes here. Is there a performance improvement? If so, is there a performance comparison that you've done compared to the previous iteration?

Apologies for getting back late here on my side!