@samukweku 2024-12-01
@ericmjl i have updated the example to show the perf. improvement